requirements.txt
langchain_openai
langchain_core
langchain_community
python-dotenv
langchain_ollama
streamlit
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
arxiv
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
langchainhub
sentence-transformers
tf-keras
langchain-huggingface
transformers
pinecone
pinecone-client
pinecone-text
pinecone-notebooks
